// src/shared/api/generated/models/examinationManageContractDtoUserUserOrderBase.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * i车保WebApi中间件
 * OpenAPI spec version: v1
 */
import type { ExaminationManageContractDtoUserUserOrderBaseState } from './examinationManageContractDtoUserUserOrderBaseState';

export interface ExaminationManageContractDtoUserUserOrderBase {
  id?: string;
  examTypeId?: string;
  examTypeName?: string;
  examTypeCode?: string;
  examTypePrtId?: string;
  examTypePrtCode?: string;
  userId?: string;
  /** 会员规则Id */
  vipRuleId?: string;
  /** 会员规则时长 */
  month?: number;
  /** 是否是Vip */
  readonly isVip?: boolean;
  /** 状态 */
  state?: ExaminationManageContractDtoUserUserOrderBaseState;
  /** 状态 */
  readonly stateText?: string;
  /** 原价 */
  originalAmount?: number;
  /** 支付金额 */
  amount?: number;
  /** 输入的好友邀请码 */
  iptInviteCode?: string;
  /** 创建时间 */
  createTime?: string;
  /** 过期时间 */
  expirationTime?: string;
  /** 过期时间（月份） */
  expirationMonth?: number;
  /** 用户设置的距离考试时间 */
  examTime?: string;
  /** 体验 次数 */
  tryOn?: number;
  /** 好友ID */
  friendId?: string;
  /** 默认 */
  isDefault?: boolean;
  /** 订单编号 */
  orderNumber?: string;
  /** 是否分享过 */
  isShare?: boolean;
  /** 是否已经使用过分享 */
  isUseInviteCode?: boolean;
}